Lock CE storage and reset strong auth flags on stopping private profiles
Always immediately clear credential-encrypted (CE) keys to lock private
profiles as soon as it is stopped. In turn, this will require strong
authentication, that is PIN or password, to unlock private space
profiles again.

This new behavior applies to existing private space, for default
UserProperties set on user types are not persisted to storage.